add feegrant msg to grant-oracle cmd (#49)
* add feegrant msg to grant-oracle cmd

* add disable delete future withdrawals flag (#50)

* add disable delete future withdrawals flag

* format

+++ provider/child/child.go
@@ -107,6 +107,7 @@ func (b *BaseChild) Initialize(
 	bridgeInfo ophosttypes.QueryBridgeResponse,
 	keyringConfig *btypes.KeyringConfig,
 	oracleKeyringConfig *btypes.KeyringConfig,
+	disableDeleteFutureWithdrawals bool,
 ) (uint64, error) {
 	b.SetBridgeInfo(bridgeInfo)
 
@@ -117,14 +118,16 @@ func (b *BaseChild) Initialize(
 
 	var l2Sequence uint64
 	if b.node.HeightInitialized() {
-		l2Sequence, err = b.QueryNextL2Sequence(ctx, processedHeight)
-		if err != nil {
-			return 0, err
-		}
+		if !disableDeleteFutureWithdrawals {
+			l2Sequence, err = b.QueryNextL2Sequence(ctx, processedHeight)
+			if err != nil {
+				return 0, err
+			}
 
-		err = b.mk.DeleteFutureFinalizedTrees(l2Sequence)
-		if err != nil {
-			return 0, err
+			err = b.mk.DeleteFutureFinalizedTrees(l2Sequence)
+			if err != nil {
+				return 0, err
+			}
 		}
 
 		version := types.MustInt64ToUint64(processedHeight)
